Removed commented out log message, changed Template to use int instead of Integer (which resulted in removing a test - since java will not allow ints to be null)

+++ src/test/java/io/lampada/template/TemplateTests.java
@@ -18,7 +18,7 @@ public class TemplateTests {
     public void getIdShouldReturnId() {
         testTemplate.setId(1);
 
-        Assert.assertEquals("Id was not 1", (Integer) 1, testTemplate.getId());
+        Assert.assertEquals("Id was not 1", 1, testTemplate.getId());
     }
 
     @Test
